[Modify] diet 화면들 screen 패키지 안으로 정리
