+++ app/src/main/java/com/konkuk/kuit_kac/presentation/navigation/KacNavGraph.kt
@@ -8,12 +8,12 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.konkuk.kuit_kac.presentation.diet.FastingResultScreen
-import com.konkuk.kuit_kac.presentation.diet.MealEditResultScreen
-import com.konkuk.kuit_kac.presentation.diet.MealMainScreen
-import com.konkuk.kuit_kac.presentation.diet.MealRecordScreen
-import com.konkuk.kuit_kac.presentation.diet.MealSearchItemDetailScreen
-import com.konkuk.kuit_kac.presentation.diet.MealSearchScreen
+import com.konkuk.kuit_kac.presentation.diet.screen.FastingResultScreen
+import com.konkuk.kuit_kac.presentation.diet.screen.MealEditResultScreen
+import com.konkuk.kuit_kac.presentation.diet.screen.MealMainScreen
+import com.konkuk.kuit_kac.presentation.diet.screen.MealRecordScreen
+import com.konkuk.kuit_kac.presentation.diet.screen.MealSearchItemDetailScreen
+import com.konkuk.kuit_kac.presentation.diet.screen.MealSearchScreen
 import com.konkuk.kuit_kac.presentation.diet.component.DietPatchScreen
 import com.konkuk.kuit_kac.presentation.diet.screen.DietCreateScreen
 import com.konkuk.kuit_kac.presentation.diet.screen.DietExistScreen
